New update!!!(Tried to make simple smart responsive page.)
I think this will become my last attempt to make responsiveness. To make this responsive I sacrifice my valuable time,data...etc. I don't know how it will impact to me. But I think this will be a good investment for my future.

And dear contributors, This is the your time to make changes if there are some errors or inappropriate things. 
                                                                                                                                                      All the best!!!
                                                                                                                                                 W.A.N.M.Ubeysinghe.